requirements.txt
anyio==4.4.0
certifi==2024.7.4
dadata==21.10.1
et-xmlfile==1.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openpyxl==3.1.5
python-dotenv==1.0.1
sniffio==1.3.1
